refactoring profile code
