+++ client/src/components/organisms/ProfileModal/ProfileModal.js
@@ -1,190 +1,61 @@
-import React, { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useState } from 'react';
 import {
   Box,
-  Button,
-  Checkbox,
-  FormControl,
-  FormLabel,
   Icon,
-  Image,
-  Input,
   Modal,
   ModalCloseButton,
   ModalContent,
-  ModalFooter,
-  ModalHeader,
   ModalOverlay,
-  Select,
   Stack,
-  Text,
   VStack,
 } from '@chakra-ui/react';
 import { FaUser, FaSchool, FaLock } from 'react-icons/fa';
-import { updateUser } from '../../../redux/userSlice';
-import { schoolsWithCoursesSelector } from '../../../redux/selectors/schoolsWithCoursesSelector';
-import { userFavouriteCoursesSelector } from '../../../redux/selectors/userFavouriteCoursesSelector';
-import { userSchoolWithCoursesSelector } from '../../../redux/selectors/userSchoolWithCoursesSelector';
+import ProfileUserSettings from '../../molecules/ProfileSettings/ProfileUserSettings/ProfileUserSettings';
+import ProfileSchoolSettings from '../../molecules/ProfileSettings/ProfileSchoolSettings/ProfileSchoolSettings';
 
 const ProfileModal = ({ isOpen, handleClose }) => {
-  const dispatch = useDispatch();
-  const user = useSelector(state => state.user);
-  const schoolsWithCourses = useSelector(schoolsWithCoursesSelector);
-  const userFavoriteCourses = useSelector(userFavouriteCoursesSelector);
-  const userSchool = useSelector(userSchoolWithCoursesSelector);
-
-  const [userInfo, setUserInfo] = useState({
-    email: user.email,
-    //password: user.password,
-    firstName: user.firstName,
-    lastName: user.lastName,
-    type: user.type,
-  });
-
   const [selectedSetting, setSelectedSetting] = useState('Personal');
-  const [selectedSchool, setSelectedSchool] = useState(userSchool);
-  const [selectedCourses, setSelectedCourses] = useState(userFavoriteCourses);
 
-  const handleSchoolChange = e => {
-    setSelectedSchool(schoolsWithCourses[e.target.value]);
-  };
-
-  useEffect(() => {
-    if (selectedSchool && userSchool && selectedSchool._id === userSchool._id) {
-      setSelectedCourses(userFavoriteCourses);
-    } else {
-      setSelectedCourses({});
-    }
-  }, [selectedSchool]);
-
-  const handleCourseChange = course => {
-    setSelectedCourses(prevCourses => {
-      if (prevCourses[course._id]) {
-        const { [course._id]: deletedCourse, ...remainingCourses } =
-          prevCourses;
-        return remainingCourses;
-      } else {
-        return { ...prevCourses, [course._id]: course };
-      }
-    });
-  };
-
-  const handleSave = () => {
-    const favourites = Object.keys(selectedCourses);
-    const school = selectedSchool._id;
-    const updatedUser = {
-      ...userInfo,
-      school,
-      favourites,
-    };
-    dispatch(updateUser(updatedUser));
-    handleClose();
-  };
-
-  const renderSchools = () => {
-    return Object.values(schoolsWithCourses).map((school, i) => (
-      <option key={i} value={school._id}>
-        {school.name}
-      </option>
-    ));
-  };
-
-  const renderCourses = () => {
-    if (selectedSchool) {
-      return Object.values(selectedSchool.courses).map((course, i) => (
-        <Checkbox
-          key={i}
-          isChecked={!!selectedCourses[course._id]}
-          onChange={() => handleCourseChange(course)}
-        >
-          {course.courseCode}
-        </Checkbox>
-      ));
-    }
-  };
-
-  const renderUserSettings = () => {
+  const renderSettingsSidePanel = () => {
     return (
-      <Box w="600px">
-        <ModalHeader> User Settings </ModalHeader>
-        <Stack direction="row">
-          <Box w="300px">
-            <Image
-              borderRadius="full"
-              boxSize="150px"
-              src="https://soccerpointeclaire.com/wp-content/uploads/2021/06/default-profile-pic-e1513291410505.jpg"
-              alt="Profile Picture"
-            />
-          </Box>
-          <Box w="100%">
-            <Stack direction="row" spacing={4}>
-              <FormControl>
-                <FormLabel>First name</FormLabel>
-                <Input
-                  placeholder="First name"
-                  value={userInfo.firstName}
-                  onChange={e =>
-                    setUserInfo({ ...userInfo, firstName: e.target.value })
-                  }
-                />
-              </FormControl>
-              <FormControl>
-                <FormLabel>Last name</FormLabel>
-                <Input
-                  placeholder="Last name"
-                  value={userInfo.lastName}
-                  onChange={e =>
-                    setUserInfo({ ...userInfo, lastName: e.target.value })
-                  }
-                />
-              </FormControl>
-            </Stack>
-            <FormControl>
-              <FormLabel>Email address</FormLabel>
-              <Input
-                placeholder="Email address"
-                value={userInfo.email}
-                onChange={e =>
-                  setUserInfo({ ...userInfo, email: e.target.value })
-                }
-              />
-            </FormControl>
-            <FormControl>
-              <FormLabel>Account Type</FormLabel>
-              <Select
-                value={userInfo.type}
-                onChange={e =>
-                  setUserInfo({ ...userInfo, type: e.target.value })
-                }
-              >
-                <option value="Student">Student</option>
-                <option value="Professor">Teacher</option>
-                <option value="Admin">Admin</option>
-                <option value="Developer">Developer</option>
-              </Select>
-            </FormControl>
-          </Box>
-        </Stack>
-      </Box>
-    );
-  };
-
-  const renderSchoolSettings = () => {
-    return (
-      <Box w="600px">
-        <ModalHeader>School Settings</ModalHeader>
-        <FormControl>
-          <FormLabel>School</FormLabel>
-          <Select
-            placeholder="Select a school"
-            value={selectedSchool?._id}
-            onChange={handleSchoolChange}
+      <Box width="150px" borderRight="1px solid" borderColor="gray.200">
+        <VStack spacing={3} height="full" justify="space-around">
+          <Box
+            as="button"
+            w="100%"
+            flex="1"
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+            onClick={() => setSelectedSetting('Personal')}
           >
-            {renderSchools()}
-          </Select>
-        </FormControl>
-        <Text fontSize="lg">Courses</Text>
-        {renderCourses()}
+            <Icon as={FaUser} boxSize="24px" />
+          </Box>
+          <Box
+            as="button"
+            w="100%"
+            flex="1"
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+            borderTop="1px solid"
+            borderBottom="1px solid"
+            borderColor="gray.400"
+            onClick={() => setSelectedSetting('School')}
+          >
+            <Icon as={FaSchool} boxSize="24px" />
+          </Box>
+          <Box
+            as="button"
+            w="100%"
+            flex="1"
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+          >
+            <Icon as={FaLock} boxSize="24px" />
+          </Box>
+        </VStack>
       </Box>
     );
   };
@@ -192,9 +63,9 @@ const ProfileModal = ({ isOpen, handleClose }) => {
   const renderSettings = () => {
     switch (selectedSetting) {
       case 'Personal':
-        return renderUserSettings();
+        return <ProfileUserSettings handleClose={handleClose} />;
       case 'School':
-        return renderSchoolSettings();
+        return <ProfileSchoolSettings handleClose={handleClose} />;
       default:
         return null;
     }
@@ -206,54 +77,8 @@ const ProfileModal = ({ isOpen, handleClose }) => {
       <ModalContent>
         <ModalCloseButton />
         <Stack direction="row" height="400px">
-          <Box width="150px" borderRight="1px solid" borderColor="gray.200">
-            <VStack spacing={3} height="full" justify="space-around">
-              <Box
-                as="button"
-                w="100%"
-                flex="1"
-                display="flex"
-                alignItems="center"
-                justifyContent="center"
-                onClick={() => setSelectedSetting('Personal')}
-              >
-                <Icon as={FaUser} boxSize="24px" />
-              </Box>
-              <Box
-                as="button"
-                w="100%"
-                flex="1"
-                display="flex"
-                alignItems="center"
-                justifyContent="center"
-                borderTop="1px solid"
-                borderBottom="1px solid"
-                borderColor="gray.400"
-                onClick={() => setSelectedSetting('School')}
-              >
-                <Icon as={FaSchool} boxSize="24px" />
-              </Box>
-              <Box
-                as="button"
-                w="100%"
-                flex="1"
-                display="flex"
-                alignItems="center"
-                justifyContent="center"
-              >
-                <Icon as={FaLock} boxSize="24px" />
-              </Box>
-            </VStack>
-          </Box>
-          <VStack p={5}>
-            {renderSettings()}
-            <ModalFooter>
-              <Button colorScheme="blue" mr={3} onClick={handleSave}>
-                Save
-              </Button>
-              <Button onClick={handleClose}>Cancel</Button>
-            </ModalFooter>
-          </VStack>
+          {renderSettingsSidePanel()}
+          <VStack p={5}>{renderSettings()}</VStack>
         </Stack>
       </ModalContent>
     </Modal>
